relocate storage translator
Moving the translator logic (for the forth time?) to detect a namespace
for anything storage-format specific. This is to prepare for future
changes where additional files will be introduced for storage
format-specific implementation.
