+++ sphinxcontrib/confluencebuilder/translator.py
